Added Flexbox Zombies and Grid Critters courses.
No longer will you have to feel overwhelmed and refer back to a cheatsheet when doing front-end development with Flebox Zombie and Grid Critters.  

Going from Wes Bos material straight to Flexbox will put you in a category that few are in and allow you to work fairly quickly when taking mockups from your designers and implementing them into production.

Following that course with Grid Critters, you will have the tools necessary to build virtually any website you can think of from scratch.  There are some frameworks I want to add still, but need more time to look into like Tachyons.  

But I digress, 9894 is for becoming good and having an understanding at many levels, which is why we will more so focus on re-inventing the wheel in the courses I suggest, while later moving to frameworks.  For work you'll more than likely stick to popular solutions that offer very precise and composable primitives that you can build on top of. 

Which actually, you know what, leads me to one more treat.

This is an article that you will want to take a look at when you get the chance, if you’re going to end up writing LESS, SASS, Stylus or vanilla CSS at the end of the day. 

http://clubmate.fi/oocss-acss-bem-smacss-what-are-they-what-should-i-use/

I have also added that article to the list as well.  This list needs more structure, I know.  As I posted about in the last update, that will be coming.  For now, though, enjoy and keep working hard.  :)