Completed Flexbox Zombies [approved]
I had the chance to go through the entire Flexbox Zombies course by Dave Geddes yesterday and I am as of today approving it on Project 9894.

This is by far, one of the most comprehensive introductions to Flexbox that I have come across and many in industry have also given the green light.  You should be able to get to an intermediate level of understanding in Flexbox, capable of building out virtually any front-end applications that you can think of.

My only recommendation is to try to build a few applications using what you've learned.

He has another course called Grid Critters, which is paid, unfortunately.  Therefore, if it is approved, it will make it to the final 9894, but it won't be the most recommended; Meaning, I will try to find something better.  If I can't then...you'll just have to drop some money.